requirement.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
